+++ src/types.ts
@@ -12,6 +12,8 @@ import {
   optimism,
   avalanche,
   worldchain,
+  berachain,
+  unichain,
 } from "viem/chains";
 
 import type {
@@ -35,8 +37,10 @@ export type SupportedChainId =
   | typeof mainnet.id
   | typeof polygon.id
   | typeof arbitrum.id
+  | typeof unichain.id
   | typeof optimism.id
   | typeof avalanche.id
+  | typeof berachain.id
   | typeof worldchain.id;
 
 export interface EnrichLogsArgs {