+++ CHANGELOG.md
@@ -1,5 +1,26 @@
 # Changelog
 
+## [2.10.0](https://github.com/0xProject/0x-parser/compare/v2.9.0...v2.10.0) (2025-02-25)
+
+
+### Features
+
+* log warning for reverted transactions ([4d0ab07](https://github.com/0xProject/0x-parser/commit/4d0ab07afc53e39ab607f22a8cca21e976197ca2))
+
+## [2.9.0](https://github.com/0xProject/0x-parser/compare/v2.8.0...v2.9.0) (2025-02-21)
+
+
+### Features
+
+* support parsing transactions on berachain ([34900e7](https://github.com/0xProject/0x-parser/commit/34900e74ef87b1555b2078b736c5e8eac3fb40e0))
+
+## [2.8.0](https://github.com/0xProject/0x-parser/compare/v2.7.1...v2.8.0) (2025-02-12)
+
+
+### Features
+
+* unichain support ([0f49539](https://github.com/0xProject/0x-parser/commit/0f4953917e56127949abfe9a06bda8003384c7c1))
+
 ## [2.7.1](https://github.com/0xProject/0x-parser/compare/v2.7.0...v2.7.1) (2025-01-21)